♻️ Refactor browser download method into a util
