+++ lib/browser/base.js
@@ -1,16 +1,15 @@
 import * as fs from 'node:fs';
 import * as os from 'node:os';
 import * as path from 'node:path';
-import * as https from 'node:https';
 import { randomBytes } from 'node:crypto';
 import { spawn } from 'cross-spawn';
 import { rimraf } from 'rimraf';
 
 import { Connection } from '../connection.js';
-import { extract } from '../util/extract.js';
 import { getPlatform } from '../util/platform.js';
 import { DeferredPromise } from '../util/promise.js';
-import { indent, formatBytes, formatTime } from '../util/string.js';
+import { download } from '../util/download.js';
+import { indent } from '../util/string.js';
 
 export class BrowserLauncher {
   static #launchers = new Map();
@@ -29,7 +28,6 @@ export class BrowserLauncher {
   name = 'browser';
   platform = getPlatform();
   id = randomBytes(8).toString('hex');
-  log = process.stderr;
 
   width = 1280;
   height = 720;
@@ -141,67 +139,11 @@ export class BrowserLauncher {
 
   async install() {
     let executablePath = await this.getExecutablePath();
-    if (fs.existsSync(executablePath)) return executablePath;
+    if (fs.existsSync(executablePath)) return;
 
-    let downloadUrl = await this.getDownloadUrl();
-    let installPath = await this.getInstallPath();
-    let downloadPath = path.join(installPath,
-      decodeURIComponent(downloadUrl.split('/').pop())
-    );
-
-    try {
-      if (!fs.existsSync(downloadPath)) {
-        this.log.write(`Downloading ${this.name}...\n`);
-        await fs.promises.mkdir(installPath, { recursive: true });
-
-        await new Promise((resolve, reject) => https.get(downloadUrl, res => {
-          if (res.statusCode !== 200) {
-            res.resume();
-            reject(new Error(`Download failed: ${res.statusCode} - ${downloadUrl}`));
-          } else {
-            let total = parseInt(res.headers['content-length'], 10);
-            let progress = { width: 25 };
-
-            res.on('data', chunk => {
-              let amount = progress.amount = (progress.amount ?? 0) + chunk.length;
-              let ratio = amount === total ? 1 : Math.min(Math.max(amount / total, 0), 1);
-              let length = progress.length = Math.round(progress.width * ratio);
-              let elapsed = Date.now() - (progress.start ??= Date.now());
-              let eta = (ratio >= 1) ? 0 : elapsed * (total / amount - 1);
-              let percent = Math.floor(ratio * 100).toFixed(0);
-
-              if (this.log.isTTY) {
-                this.log.moveCursor(0, -1);
-                this.log.clearLine(0);
-                this.log.write(`Downloading ${this.name} ` +
-                  `[${'='.repeat(length)}${' '.repeat(progress.width - length)}] ` +
-                  `${formatBytes(amount)}/${formatBytes(total)} ` +
-                  `${percent}% ${formatTime(eta)}\n`);
-              }
-            });
-
-            res.pipe(fs.createWriteStream(downloadPath)
-              .on('finish', resolve)
-              .on('error', reject));
-          }
-        }).on('error', reject));
-      }
-
-      this.log.moveCursor?.(0, -1);
-      this.log.clearLine?.(0);
-      this.log.write(`Extracting ${this.name}...\n`);
-
-      await extract(downloadPath, installPath);
-
-      this.log.moveCursor?.(0, -1);
-      this.log.clearLine?.(0);
-      this.log.write(`Installed ${this.name}\n`);
-
-      return executablePath;
-    } finally {
-      if (fs.existsSync(downloadPath))
-        await rimraf(downloadPath);
-    }
+    let url = await this.getDownloadUrl();
+    let dir = await this.getInstallPath();
+    await download(this.name, url, dir);
   }
 
   async launch() {
@@ -228,6 +170,7 @@ export class BrowserLauncher {
       this.devtools.on('Runtime.consoleAPICalled', ({ type, args }) => {
         args = args.map(a => a.value ?? a.description);
         if (args.length) args[0] = `[${this.name}] ${args[0]}`;
+        // @todo: let reporter handle logs?
         console[type]?.(...args);
       });
     }
